change yables names and add swagger API description

+++ employee/model_views.py
@@ -1,15 +1,19 @@
 from django.core.exceptions import ObjectDoesNotExist
 from django.db import models
 from rest_framework import status, serializers, permissions
+from rest_framework.authentication import TokenAuthentication
 from rest_framework.response import Response
 from rest_framework.views import APIView
+
+from employee.permissions import IsAdminOrReadOnly
 from employee.utils import Utils
 
 
 class MultipleInstanceAPIView(APIView):
     serializer = serializers.Serializer
     model = models.Model
-    permission_classes = (permissions.IsAuthenticated,)
+    authentication_classes = (TokenAuthentication, )
+    permission_classes = (permissions.IsAuthenticated, IsAdminOrReadOnly)
 
     def get(self, request):
         models = self.model.objects.all()
@@ -26,7 +30,7 @@ class MultipleInstanceAPIView(APIView):
 class SingleInstanceAPIView(APIView):
     serializer = serializers.Serializer
     model = models.Model
-    permission_classes = (permissions.IsAuthenticated,)
+    permission_classes = (permissions.IsAuthenticated, IsAdminOrReadOnly)
 
     def get(self, request, model_id):
         try:
@@ -66,13 +70,16 @@ class MultipleEmployeeRelatedInstanceAPIView(APIView):
         return Response(self.serializer(models, many=True).data)
 
     def post(self, request, employee_id):
-        data = request.data
-        data['employeeId'] = employee_id
-        serializer = self.serializer(data=data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Utils.error_response(serializer.errors, status.HTTP_400_BAD_REQUEST)
+        if self.__owner_or_admin(request, employee_id):
+            data = request.data
+            data['employeeId'] = employee_id
+            serializer = self.serializer(data=data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Utils.error_response(serializer.errors, status.HTTP_400_BAD_REQUEST)
+        else:
+            return Utils.error_response("Permission denied", status.HTTP_403_FORBIDDEN)
 
 
 class SingleEmployeeRelatedInstanceAPIView(APIView):
@@ -90,19 +97,29 @@ class SingleEmployeeRelatedInstanceAPIView(APIView):
     def patch(self, request, employee_id, model_id):
         try:
             model = self.model.objects.get(id=model_id, employee_id=employee_id)
-            serializer = self.serializer(model, data=request.data, partial=True)
-            if serializer.is_valid():
-                serializer.save()
-                return Response(serializer.data, status.HTTP_200_OK)
+            if self.__owner_or_admin(request, employee_id):
+                serializer = self.serializer(model, data=request.data, partial=True)
+                if serializer.is_valid():
+                    serializer.save()
+                    return Response(serializer.data, status.HTTP_200_OK)
+                else:
+                    return Utils.error_response(serializer.errors, status.HTTP_400_BAD_REQUEST)
             else:
-                return Utils.error_response(serializer.errors, status.HTTP_400_BAD_REQUEST)
+                return Utils.error_response("Permission denied", status.HTTP_403_FORBIDDEN)
         except ObjectDoesNotExist as e:
             return Utils.error_response(e.args, status.HTTP_404_NOT_FOUND)
 
     def delete(self, request, employee_id, model_id):
         try:
             model = self.model.objects.get(id=model_id, employee_id=employee_id)
-            model.delete()
-            return Response({'id': model_id}, status.HTTP_200_OK)
+            if self.__owner_or_admin(request, employee_id):
+                model.delete()
+                return Response({'id': model_id}, status.HTTP_200_OK)
+            else:
+                return Utils.error_response("Permission denied", status.HTTP_403_FORBIDDEN)
         except ObjectDoesNotExist as e:
             return Utils.error_response(e.args, status.HTTP_404_NOT_FOUND)
+
+    @staticmethod
+    def __owner_or_admin(request, employee_id):
+        return employee_id == request.user or request.user.is_staff
